fix: 가족 탭 API 연동 안정화 및 버그 수정

+++ app/src/main/java/com/example/senior_on/data/remote/dto/FamilyDtos.kt
@@ -32,4 +32,4 @@ data class FamilyHomeResponse(
     val recentUploaderProfileImageUrls: List<String>?,
     val recentPhotos: List<FamilyPhotoItemResponse>?
 )
-data class FamilyCodeResponse(val familyCode: String?, val familyMemberCount: Int?)
+data class FamilyCodeResponse(val familyCode: String?, val familyMemberCount: Long?)
